+++ src/TestSuite/IntegrationTestCase.php
@@ -14,14 +14,14 @@
 namespace Thumber\TestSuite;
 
 use Cake\TestSuite\IntegrationTestCase as CakeIntegrationTestCase;
-use Thumber\ThumbTrait;
+use Thumber\ThumbsPathTrait;
 
 /**
  * IntegrationTestCaseTest class
  */
 abstract class IntegrationTestCase extends CakeIntegrationTestCase
 {
-    use ThumbTrait;
+    use ThumbsPathTrait;
 
     /**
      * Teardown any static object changes and restore them