+++ tests/TestCase/TestSuite/TestCaseTest.php
@@ -14,7 +14,7 @@ namespace Thumber\Test\TestCase\TestSuite;
 
 use Cake\Core\Configure;
 use Thumber\TestSuite\TestCase;
-use Thumber\ThumbTrait;
+use Thumber\ThumbsPathTrait;
 use Thumber\Utility\ThumbManager;
 
 /**
@@ -22,7 +22,7 @@ use Thumber\Utility\ThumbManager;
  */
 class TestCaseTest extends TestCase
 {
-    use ThumbTrait;
+    use ThumbsPathTrait;
 
     /**
      * Test for `assertImageFileEquals()` method