+++ tests/TestCase/Utility/ThumbManagerTest.php
@@ -13,7 +13,7 @@
 namespace Thumber\Test\TestCase\Utility;
 
 use Thumber\TestSuite\TestCase;
-use Thumber\ThumbTrait;
+use Thumber\ThumbsPathTrait;
 use Thumber\Utility\ThumbCreator;
 use Thumber\Utility\ThumbManager;
 
@@ -22,7 +22,7 @@ use Thumber\Utility\ThumbManager;
  */
 class ThumbManagerTest extends TestCase
 {
-    use ThumbTrait;
+    use ThumbsPathTrait;
 
     /**
      * @var \Thumber\Utility\ThumbManager