+++ CHANGELOG.md
@@ -3,10 +3,10 @@
 # 1.5.0
 * the plugin has been migrated to CakePHP 3.6;
 * added `TestCase::skipIfDriverIs()` method;
+* `ThumbTrait` becomes `ThumbsPathTrait` and it has no more `getDriver()`,
+    `getExtension()` and `getUrl()` methods;
 * the `ThumbCreator::getDefaultSaveOptions()` method now can take the `$path`
     optional argument;
-* removed `getDriver()`, `getExtension()` and `getUrl()` methods provided by
-    the `ThumbTrait`;
 * removed `ThumbTrait::getSupportedFormats()` method, added 
     `ThumbManager::$supportedFormats` static property;
 * `InternalErrorException` exception has been replaced with `RuntimeException`.